qcacld-2.0: Avoid switch channel during preauth
prima to qcacld-2.0 propagation

Currently data tx is suspended and later checked for preauth
in limProcessChannelSwitchTimeout. This can result in a condition
that data tx to be always in suspend case if roaming triggers at
the same time.

Fix is to add a check for preauth during channel switch.

Change-Id: I0ac3d63a5db000beed7de34b13e5926c7dc3ce48
CRs-Fixed: 2017768