qcacld-2.0: Pause and unpause durations configurable via ini file
The isolation between WLAN and BT antenna is not sufficient for WLAN and
BT to operate independently eventhough they have separate antenna. So, as
a workaround to reduce interference from WLAN Tx on BT Rx, WLAN on the uplink
periodically pauses and unpauses its data queues,configurable via gCoexWLANPauseDuration
and gCoexWLANUnpauseDuration in ini file.

Change-Id: I4f3c9227bdd2f1352f0209683f3ad9aea5c52473
CRs-Fixed: 2043806