qcacld-2.0: Wait for the completion while cleaning up the SAP session
Sometimes the mode change requests received in quick succession
while the previous session is not opened. It leads to failure
when all the sessions are allocated and the requested one has
to be recycled for use.

To mitigate this, wait for the completion while cleaning up the SAP
session.

Change-Id: Ifaaefec7370edcf0eee72e30e020bb63f22fdced
CRs-Fixed: 2031881