qcacld-2.0: Reduce rssi margin for Japan bin5 pulses
Japan bin5 pulses reported to host are consistently in
25-26 range even with a tuned BDF. This BDF yields good
performance across all other domain and channels.
With margin 5, the effective rssi threshold is 27 and
the bin5 radar pulses are rejected in host by a margin
of 2 dbm. so the rssi margin is reduced based on observation
to improve the detection.

CRs-Fixed: 2121088
Change-Id: I8c614cda05d7a972ccbce436d6d4c0001e35f336